Fix unit tests

+++ kong/route_service_test.go
@@ -81,7 +81,7 @@ func TestRoutesRoute(T *testing.T) {
 	assert.Equal("snihost1.com", *createdRoute.SNIs[0])
 	assert.Equal("snihost2.com", *createdRoute.SNIs[1])
 	assert.Equal("10.0.0.0/8", *createdRoute.Destinations[0].IP)
-	assert.Equal(80, *createdRoute.Destinations[0].Port)
+	assert.Equal(int64(80), *createdRoute.Destinations[0].Port)
 
 	err = client.Routes.Delete(defaultCtx, createdRoute.ID)
 	assert.Nil(err)
